Safari-609.1.11.1/LayoutTests/js/dfg-mod-1-int-expected.txt

Tests that the DFG handles x % 1, where x is an integer, correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(1) is 0 on all iterations including after DFG tier-up.
PASS successfullyParsed is true

TEST COMPLETE
